+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+- Add Rector rule set to update the hook_event_dispatcher module
+
+### Changed
+- Update dependency injection Rector rule set
+- Remove ChangeOrIfReturnToEarlyReturnRector Rector rule
+
 ## [1.6.0] - 2021-03-12
 ### Added
 - Add PHP 8 support